btf: Add split BTF support.
Split BTFs are built in a way that their type ids and string offsets are
consecutive. So when parsing the raw object we just have to correctly
offset those values.

The Btf object takes care of querying the base if present and keep the
API consistent while hiding the presense of a base. Additionally, by
using a reference counted pointer, the data from the base is reused.

For unit test, add a test module (openvswitch) from the same kernel
version as the current vmlinux: 5.17.4-200.fc35.x86_64.

Signed-off-by: Adrián Moreno <amorenoz@redhat.com>