tree: lookups by name aren't guaranteed to be unique
In a base BTF or single split + base BTF, ids are guaranteed to be
unique. But this is not the case for the strings, which can be used by
more than one type.

Fix this by making a breaking API change and returning a vec instead of
a single entry for "by name" lookups.

Signed-off-by: Antoine Tenart <atenart@redhat.com>