btf.rs: return the correct type for Array
as per BTF specification, btf_type is followed by btf_array info that
describes:

- element type
- index type
- number of elements

the parsing part of the library extracts such info, but the type id
getter still refers to btf_type (which is required to be zeroed except
for the kind field) for the type_id.

Return the correct member describing the type id.

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>