tree: stop returning an error for empty resolve results
Before this patch, resolve helpers were all returning a Result and an
error when the returned value was empty. This did not map well to what
should be a Result and did not allow to match actual errors (if any).

This moves the resolve helpers to either return a simple Vec (in case
the lookup can't fail) or a Result<Vec> it can it can. Empty result are
no longer converted to an error. In cases where a single result can be
returned it is now wrapped in an Option instead of an Error.

There are two kinds of failures in resolve helpers:

- The BtfType helper is not defined. Calling it is against the BTF spec
  (eg. there is no linked type defined for that BTF type).
- We do a chained lookup and the second one fails even though the input
  data comes from our BTF object (eg. we lookup ids using a name, and
  then types using those ids).

This of course breaks the API.

Signed-off-by: Antoine Tenart <atenart@redhat.com>