btf: make Type non-exhaustive
Even though this doesn't happen too often, BTF types can be added upstream. We
don't want to have to release a major version for those kind of changes, and by
design the BTF types are not an exhaustive list at any moment in time. Mark them
as non-exhaustive to ease support and usage.

Signed-off-by: Antoine Tenart <atenart@redhat.com>