cbtf/btf: stop special casing anonymous type names
Anonymous type names where special cased because there was no proper way of
describing if it was valid on a per-type basis. This is not longer true with the
addition of BtfKind as we can now filter at the cbtf level.

This removes the special ANON_TYPE_NAME definition and moves the anonymous name
check in the cbtf part. This is more logical as the validity of an anonymous
name is defined in the BTF format definition; cbtf types now return a name
offset, including 0, only when valid.

This makes things better scoped and helps with the following:

- Allows the support of the mmap backend later.
- Covers the anonymous type support in all the public APIs (regex helpers and
  resolve_ids_by_name).

Also improved the anonymous test cases to cover all API helpers including regex.

Signed-off-by: Antoine Tenart <atenart@redhat.com>